Added macros MAP_UD, MAP_LIST_UD, MAP_UD_I, MAP_LIST_UD_I.
These macros allow the user to pass a single value to all invocations of the user macro,
and the *_I variants additionally pass the index of the invocation as an integer literal.
This requires a list of predefined increment operations. Updated the documentation
accordingly, and added an overview and a link to the µSer library to README.md.